Finalize tree checkbox API (#185846)
* Finalize tree checkbox API
Fixes #116141

* Remove TODO
